WIP: Implemented SDP mangling for recvonly answer
AND-232

Implemented SDP mangling for recvonly sdp answer. Seems to work but need more testing and clean up.
The webrtc designed behaviour is that if an offerer SDP indicates to send media, the answerer will generate an SDP to accept it, even if the answerer had put in its media constraint not to accept that media:
https://code.google.com/p/webrtc/issues/detail?id=2404
Hence, for our answerer to respect its own MediaConstraints, the SDP will be mangled if needed to respect the MediaConstraints (sdpMediaConstraints).

Major changes:
- Mangle SDP to respect MediaConstraints as an answerer.
  - Mangle at sendLocalDescription if type is answer.