ENH: Added config for receiving audio and video
AND-232

Fully implemented separate config for receiving audio and video.
Not handle Peer's Audio or video tracks manually (at onAddStream) to satisfy hasVideoReceive() and hasAudioReceive(), as webrtc sdp will take care of it, albeit possibly with us doing SDP mangling when we are the answerer.
Note that PeerConnectionFactory.initializeAndroidGlobals still always use true for initializeAudio and initializeVideo, as otherwise, new PeerConnectionFactory() crashes, even if sending of audio and video are not enabled.
However, when not sending audio and/or video, their respective Android permissions (as given below) were tested to be not required:
For audio: android.permission.RECORD_AUDIO
For video: android.permission.CAMERA
)

Major changes:
- Corrected logic of mangling sdp to be at onCreateSuccess of answer SDP:
- Corrected error in sdpReplace:
  - Added \r\n to every line.
- Added sdpAudioSendOnly and sdpVideoSendOnly and corrected to use these instead of recvonly methods
  - Changed sdpReceiveOnly to sdpSendReceiveOnly to handle both send and recv only.
- Corrected to check audio and video send settings before mangling SDP to send only.
- PeerConnectionFactory.initializeAndroidGlobals to always use true for initializeAudio and initializeVideo.
- Not handle Peer's Audio or video tracks manually (at onAddStream) to satisfy hasVideoReceive() and hasAudioReceive().