WIP: Discovered cause of problem, to apply fix.
AND 304

Failure to reconnect was due to the SDK calling to disconnect the Socket connection when Socket.io creates a websocket error due to disruption of internet connection.
Removing the call to disconnect due to error would allow Socket.io to reconnect.
To apply reconnection strategy next (likely via restart to have a clean Skylink SDK state).