BUG: Fixed reconnection after disconnection.
AND 304

Failure to reconnect was due to the SDK calling to disconnectFromRoom() when Socket.io creates a websocket error due to disruption of internet connection. disconnectFromRoom() resulted in setting to null of various important elements,
such as localMediaStream, peerConnectionFactory, etc.
Fixed the error and allowed to reconnect to room in a clean Skylink SDK state.
Further refactored connection functions to SkylinkConnectionService.

Major changes:
- Used same logic for first connection and reconnection to room.
  - On disconnect from socket.io, bye will be sent by signaling server.
  - Hence, we should remove all prior Peer related info (e.g. PC, DC),
  - and enter room again with fresh socket id (i.e. PeerId).
- Moved restart functions to SkylinkConnectionService.

Testing done:
- Ran unit tests via Gradle script:
  - ./gradlew clean test
- Tested with MultiPartyVideoCall sample app.