ENH: Fixed public P2P msg sending with MCU
AND 316

Unable to send as error checking stopped sending of DC message when null userData of MCU Peer was obtained.
Fixed sending of public DC message.
Further fixed to ensure public P2P messages are not sent to MCU (as a Peer).
Still needed to have audio and video to connect as a chat client.

Major changes:
- Do not send message to MCU as a peer.
- UserInfo added empty string as default userData.
- Avoided sending 1 extra public DC message (to the MCU) when in MCU room.
- Add room names in Constants.java in Sample app in order to make changing of room names during debugging much easier.

Testing done:
- Ran unit tests via Gradle script:
  - ./gradlew clean test
- Tested with MultiPartyVideoCall sample app.
- Tested with Chat sample app.