ENH: Disabled data transfer for MCU.
AND 320

Added usage of MCU DC for binary data sending when in MCU room, but disabled for this release,
as MCU does not currently properly handle binary sending.
Hence in MCU room, binary sending will result in no action, and an exception will be thrown.
Tested to work as expected for both rooms with and without MCU.

Major changes:
- In MCU room, send Public DATA to MCU DC ONLY, and not to any Peer DC.
- Add usage of MCU DC if in MCU room.
  - Use MCU DC for send operation.
  - Use Peer DC receive operation.
- Refactored sending of binary data.
- Added hack to not support sending of binary data when in MCU room.

Testing done:
- Ran unit tests via Gradle script:
  - ./gradlew clean test
- Tested with Data sample app.