ENH: Self video mirrored when SkylinkConfig set
AND-375

Self video from front camera is rendered as NOT mirrored by default.
New config in SkylinkConfig to change mirroring for local video of front camera.
To allow the back camera to stay unmirrored, a new renderer with a new GLSurfaceView has to be created, each time the camera switches.
This means the SDK user has to add a new local video view (from onLocalMediaCapture) each time the camera is switched.

Major changes:
- Created boolean mirrorLocalView in SkylinkConfig
  - Default value is false.
  - Accessible by public getter and setter:
    - isMirrorLocalView
    - setMirrorLocalView
- Refactored getting of VideoView from VideoTrack into 2 methods:
  - genVideoViewFromVideoTrack to simply create and return a videoView from VideoTrack.
  - createVideoView to create new videoView (via genVideoViewFromVideoTrack) with logic to control if mirroring is applied.
- Removed SkylinkConfig copy constructor.
  - Changed SkylinkConnection.init to use the SkylinkConfig passed in as parameter as the SkylinkConnection.SkylinkConfig.
  - Change the checks on videoWidth, videoHeight, videoFps to their setters.
- Changed tests to test logic of switching camera only if at least 2 cameras are available.

Testing done:
- Ran unit tests via Gradle script:
  - ./gradlew clean test
- Tested with Video sample app.
- Tested with MultiPartyVideoCall sample app.